pyproject.toml
[project]
name = "syrin"
version = "0.11.0"
description = "Python library for building AI agents with budget management, declarative definitions, and production-ready observability."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Syrin Labs" }]
keywords = ["ai", "agents", "llm", "budget"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "chevron>=0.14.0",
    "pydantic>=2.0",
    "httpx",
    "tiktoken",
    "python-dotenv>=1.2.2",
    "jsonschema>=4.0",
    "croniter>=6.2.2",
    "requests>=2.33.0", # CVE-2026-25645 fix (transitive via tiktoken)
    "rich>=13.0",
    "opentelemetry-proto>=1.20.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
    "opentelemetry-exporter-otlp-proto-common>=1.20.0",
    "pygments>=2.20.0",
]

[project.urls]
Homepage = "https://syrin.ai/"
"GitHub" = "https://github.com/syrin-labs/syrin-python"
"Documentation" = "https://syrin.ai/docs"
"PyPI" = "https://pypi.org/project/syrin/"

[project.optional-dependencies]
# LLM Providers
openai = ["openai>=2.26.0"]
anthropic = ["anthropic"]
litellm = ["litellm"]

# HTTP serving (FastAPI + Uvicorn)
serve = ["fastapi>=0.134.0", "uvicorn>=0.41.0", "python-multipart>=0.0.22"]

# Knowledge & Memory Backends
vector = ["qdrant-client", "chromadb", "redis"]
postgres = ["asyncpg>=0.29", "pgvector>=0.3"]
sqlite = ["sqlite-vec>=0.1"]
knowledge = ["chonkie>=1.0", "flatbuffers"]

# Generation & Voice
generation = ["google-genai"]
voice = ["elevenlabs>=1.0", "deepgram-sdk>=3.0", "cartesia>=1.0"]

# Documents
pdf = ["docling>=2.70", "pandas", "tabulate"]
pdf-output = ["weasyprint>=60.0"]
docling = ["docling>=2.70", "pandas", "tabulate"]
docx = ["python-docx>=1.0.0"]
excel = ["openpyxl>=3.1.0"]

# Development
dev = ["pytest", "pytest-asyncio", "pytest-cov", "ruff", "mypy", "numpy"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/syrin"]

[tool.hatch.build.targets.sdist]
# Preserve src/ so wheel-from-sdist finds src/syrin (wheel expects packages path).
only-include = ["src"]
# Ensure playground_static (built by CI before build) is in sdist.
force-include = { "src/syrin/serve/playground_static" = "src/syrin/serve/playground_static" }

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests", "examples"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
# E501: line length (handled by formatter); E402: module level import after code;
# ARG002: self/cls in methods; ARG001: unused function args (e.g. pytest fixtures);
# F821: undefined name (dynamic); F811: reimport (conditional provider imports).
ignore = ["E501", "E402", "ARG002", "ARG001", "F821", "F811"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_any_explicit = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "examples.*"
disallow_untyped_defs = false
check_untyped_defs = false

[[tool.mypy.overrides]]
module = "anthropic"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "litellm"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "syrin.providers.anthropic"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "syrin.providers.openai"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "syrin.providers.litellm"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "deepgram"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["elevenlabs", "elevenlabs.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "cartesia"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "phoenix.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "langfuse"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "opentelemetry.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "qdrant_client"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "qdrant_client.models"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "chromadb"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sentence_transformers"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "chromadb.config"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "docling",
    "docling.document_converter",
    "docx",
    "openpyxl",
    "weasyprint",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "redis"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "syrin.memory.backends.redis"
disable_error_code = ["assignment", "misc"]

[[tool.mypy.overrides]]
module = "psycopg2"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "fastapi"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "starlette"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "jsonschema"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "chevron"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "chonkie"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "openai"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "numpy"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["google", "google.genai"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["asyncpg", "pgvector", "pgvector.asyncpg", "sqlite_vec"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "syrin.knowledge.stores._postgres",
    "syrin.knowledge.stores._qdrant",
    "syrin.knowledge.stores._chroma",
    "syrin.knowledge.stores._sqlite",
]
disable_error_code = [
    "union-attr",
    "has-type",
    "arg-type",
    "misc",
    "assignment",
    "no-redef",
    "attr-defined",
    "name-defined",
    "call-overload",
    "no-any-return",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short -W ignore::ResourceWarning"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore:@tool.*has no description:UserWarning",
]

[tool.coverage.run]
source = ["src/syrin"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pip-audit>=2.10.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
